Bug Fixes and New English Translate
Changed it so that it doesn't give the sentence in tokens and then the
english translation, but instead it gives out each 1 token and its
translated version per line. Also added a English translation so that we
don't have to run it through a different code to translate it back to
english. The English translator translates each token to english, and if
it can't then it prints out which token it can't translate.